Add Game, Gesture, and Player class skeleton
WIP The Game, Gesture, and Player Class.
Game will run the game and have the main Game Loop for player and a run loop for starting the game and playing it again if user wants to.
Gesture and Player class will be abstract.
Player will be inherited by a Human and Computer class.
Each RPSLS will inherit from Gesture.